chore: Update user ID in SlackEventHandlers.publish_app_home()
This commit updates the user ID parameter in the publish_app_home() method of the SlackEventHandlers class. Instead of using the "event.user" property, it now uses the "user.id" property to register new users in the database. This change ensures that the correct user ID is used for registration.